Guard browser storage access with tap to unlock wallet
We first check if there is a logged in user. If there is no
logged in user we do not check the browser storage access since
the user has not interacted with deso identity yet.

If there is a logged in user, we check the browser support and
decide whether to display the deso identity iframe message for
"tap to unlock your wallet"

If we have not previously checked the browser support due to no
user being logged in, we check when the user attempts to log in.

Also provides a hook/callback that can be passed when deso identity is
initialized that gets called once deso identity is ready.